Security/requirements.txt
mysql-connector-python
watchdog
python-magic
requests
numpy
scikit-learn